Semaphore: remove 16-bit limit
Semaphore had a historical 16-bit maximum count limit - this was odd, as
the initial count was 32-bit.

Remove the limitation - change constructors to take uint32_t, and make
the default limit 0xffffffff rather than 0xffff.

Limits of overload resolution prevent adding new constructors while
retaining the old ones.

Addresses issue raised in #10216.

+++ rtos/Semaphore.h
@@ -49,7 +49,7 @@ public:
 
       @note You cannot call this function from ISR context.
     */
-    Semaphore(int32_t count = 0);
+    Semaphore(uint32_t count = 0);
 
     /** Create and Initialize a Semaphore object used for managing resources.
       @param  count     number of available resources
@@ -57,7 +57,7 @@ public:
 
       @note You cannot call this function from ISR context.
     */
-    Semaphore(int32_t count, uint16_t max_count);
+    Semaphore(uint32_t count, uint32_t max_count);
 
     /** Wait until a Semaphore resource becomes available.
       @param   millisec  timeout value. (default: osWaitForever).
@@ -96,7 +96,7 @@ public:
     ~Semaphore();
 
 private:
-    void constructor(int32_t count, uint16_t max_count);
+    void constructor(uint32_t count, uint32_t max_count);
 
     osSemaphoreId_t               _id;
     mbed_rtos_storage_semaphore_t _obj_mem;
